+++ lib/src/utils/widget_utils.dart
@@ -1,8 +1,9 @@
-import 'package:analyzer/dart/element/element.dart';
+import 'package:analyzer/dart/element/element2.dart';
 import 'package:analyzer/dart/element/type.dart';
+import 'package:klin_dart/src/utils/constants.dart';
 
-/// Extensions for widget-related checks on ClassElement
-extension WidgetClassChecks on ClassElement {
+/// Extensions for widget-related checks on ClassElement2
+extension WidgetClassChecks on ClassElement2 {
   /// Checks if this class is a widget
   bool isWidget() {
     // Direct check for core Flutter widgets
@@ -22,8 +23,8 @@ extension WidgetClassChecks on ClassElement {
 
     // Check superclass
     final supertype = this.supertype;
-    if (supertype != null && supertype.element is ClassElement) {
-      return (supertype.element as ClassElement).isWidget();
+    if (supertype != null && supertype.element3 is ClassElement2) {
+      return (supertype.element3 as ClassElement2).isWidget();
     }
 
     return false;
@@ -48,9 +49,9 @@ extension WidgetClassChecks on ClassElement {
       'ConsumerStatefulWidget',
     ];
 
-    if (widgetBaseClasses.contains(name)) {
-      final libraryPath = library.source.uri.toString();
-      return libraryPath.contains('flutter');
+    if (widgetBaseClasses.contains(name3)) {
+      final libraryPath = library2.uri.toString();
+      return libraryPath.contains(Constants.flutter);
     }
 
     return false;
@@ -58,19 +59,21 @@ extension WidgetClassChecks on ClassElement {
 
   /// Checks if this class has a build method that takes a BuildContext
   bool hasBuildMethod() {
-    return methods.any((method) =>
-        method.name == 'build' &&
-        method.parameters[0].type
-            .getDisplayString(withNullability: false)
-            .contains('BuildContext'));
+    return methods2.any((method) =>
+        method.name3 == Constants.build &&
+        method.formalParameters[0].type
+            .getDisplayString()
+            .contains(Constants.buildContext));
   }
 
   /// Checks if this class has a createState method that returns a State with a build method
   bool hasCreateStateMethod() {
-    return methods.any((method) =>
-        method.name == 'createState' &&
-        method.returnType.element is ClassElement &&
-        (method.returnType.element as ClassElement).hasBuildMethod());
+    return methods2.any(
+      (method) =>
+          method.name3 == Constants.createState &&
+          method.returnType.element3 is ClassElement2 &&
+          (method.returnType.element3 as ClassElement2).hasBuildMethod(),
+    );
   }
 }
 
@@ -78,8 +81,8 @@ extension WidgetClassChecks on ClassElement {
 extension WidgetTypeChecks on DartType {
   /// Checks if this type is a widget type
   bool isWidget() {
-    final element = this.element;
-    if (element is ClassElement) {
+    final element = element3;
+    if (element is ClassElement2) {
       return element.isWidget();
     }
     return false;